fix: bags after refactor, adaptive Flat page
